requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --find-links=.
#
--find-links https://felipefg.github.io/pip-wheels-riscv/wheels/

certifi==2023.7.22
    # via requests
charset-normalizer==3.2.0
    # via requests
idna==3.4
    # via requests
numpy==1.24.3
    # via
    #   -r requirements.in
    #   tflite-runtime
pillow==9.5.0
    # via -r requirements.in
requests==2.31.0
    # via -r requirements.in
tflite-runtime==2.12.0
    # via -r requirements.in
urllib3==2.0.4
    # via requests
